Don't locally commit changes in multiplayer mode prior to transaction resolution
In the interests of snappy responsiveness we used to commit picks locally before getting confirmation that they had been committed to firestore. This left us open to various inconsistent states, and on occastion even corrupted drafts (multiple picks from same pack, etc.).

With this commit, we only commit locally in multiplayer mode when we get an updated snapshot from firestore. This means that the user only ever sees and interacts with valid known firestore states.

We also introduce a glass effect over the UI which is applied while we wait for the transaction to complete (this is so the user can't further interact with the UI in an uncommitted state).

This does introduce some small lag on picks (~ 200-500ms) but at the sametime eliminates the possibility of some maddeningly difficult to diagnose bugs/states.

If users find this lag intolerable, we could consider optmistically writing the change locally but then glassing over the UI (so no further changes are possible). Not sure though if this would be a slippery slope back to bug town...